refactor: update the file name to match pattern
